@@ -14,9 +14,9 @@ packages = ["claude_code_proxy"]
 [project]
 name = "claude-code-proxy-api"
 version = "0.1.0"
-description = "Add your description here"
+description = "API server that provides an Anthropic and OpenAI compatible interface over Claude Code, allowing to use your Claude OAuth account or over the API."
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 dependencies = [
   "anthropic>=0.57.1",
   "claude-code-sdk>=0.0.14",     # Using GitHub version with PR #60 fix
@@ -27,6 +27,7 @@ dependencies = [
   "pydantic-settings>=2.4.0",
   "rich>=13.0.0",
   "typer>=0.16.0",
+  "typing-extensions>=4.0.0",
   "uvicorn>=0.34.0",
 ]
 
